requirements.txt
tensorflow
tensorflow_hub
Pillow
flask
flask_restful
flask_cors